Fix Doctrine console
When working on the standalone version of the library, the Doctrine
command line utility in vendor/bin/ failed because its configuration in
cli-config.php tried to include a database configuration outside of the
FundraisingStore directory.

Changed the config name in .gitignore which still contained an ini file
instead of the php file.