Fixed bug in Subscription::isUnconfirmed
As identified by @gbirke in https://github.com/wmde/FundraisingStore/pull/107

I did not go for the bitfield approach.

Bitfields are hard to understand and IMO
should not be used unless there is really good reason for such optimization.
And when that is done, it should be hidden as well as possible.

The more concrete reason which made me decide to change this now is that
bitfields imply you can have many of the possible values. AFAIK this is not
the case here. Does a combination of "confirmed" and "needs moderation"
make any sense? AFAIK not.