Use associations for donation payment entity
Payment is now an abstract class, with one implementation -
SofortPayment. Other payments will still use the legacy method of
storing the payment data in the data blob and setting
Donation::paymentType.

Future versions of the store may move the other payment data out of the
data blob until the paymentType property of Donation can be dropped.